Add waf build script
This build script will attempt to check that the kernel configuration
is correct, that 'perf' is patched, and that necessary libraries are
available for installation.

It will install by default to PREFIX=${PWD}/install, rather than /.

It handles header dependencies correctly, so will rebuild the program
if any of its dependencies are modified.

Debug mode can be enabled by specifying --debug when running configure,
which will also turn off optimization.

The CPU model and family are shown, and the microarchitecture is guessed
where possible. If the CPU model == 42, a workaround is applied to the
event listing to set precise=0.

The relevant script for the current microarchitecture, as well as its
listing of events are installed to ${PREFIX}/bin. It is named
gooda-collect. If necessary, the scripts are modified to hard-code the
PERF path which can be overwritten at configure time with:

    PERF=/path/to/perf ./waf configure

The whole shebang can be run by saying:

    ./waf go